Consolidate build_tree creation.
