+++ open_ce/get_graph.py
@@ -19,22 +19,22 @@
 
 import os
 
-from open_ce.inputs import Argument
+from open_ce.inputs import Argument, ENV_BUILD_ARGS
 from open_ce import graph
-from open_ce import build_env
+from open_ce.build_tree import construct_build_tree
 from open_ce import utils
 
 COMMAND = "graph"
 
 DESCRIPTION = 'Plot a dependency graph.'
 
-ARGUMENTS = build_env.ARGUMENTS + \
+ARGUMENTS = ENV_BUILD_ARGS + \
             [Argument.WIDTH,
              Argument.HEIGHT]
 
 def export_graph(args):
     '''Entry Function'''
-    build_tree = build_env.construct_build_tree(args)
+    build_tree = construct_build_tree(args)
     os.makedirs(args.output_folder, exist_ok=True)
     graph_output = os.path.join(args.output_folder, utils.DEFAULT_GRAPH_FILE)
     graph.export_image(build_tree._tree, graph_output, args.width, args.height) # pylint: disable=protected-access
